Добавлен пример решающий ошибку

+++ src/Test_workshop/boxingUnboxing/Program.cs
@@ -1,10 +1,11 @@
 ﻿
 
 
+//object box = (int)42;
+//long unbox = (long)box;
+
 object box = (int)42;
-long unbox = (long)box;
-
-
+long unbox = (long)(int)box;
 
 
 Console.ReadLine();
